+++ core/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.33.1] - 2024-12-20
+
 ### Changed
 
 - Disable `default_features`
@@ -450,7 +452,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#61]: https://github.com/dusk-network/phoenix/issues/61
 
 <!-- VERSIONS -->
-[Unreleased]: https://github.com/dusk-network/phoenix/compare/v0.33.0...HEAD
+[Unreleased]: https://github.com/dusk-network/phoenix/compare/v0.33.1...HEAD
+[0.33.1]: https://github.com/dusk-network/phoenix/compare/v0.33.0...v0.33.1
 [0.33.0]: https://github.com/dusk-network/phoenix/compare/v0.32.0...v0.33.0
 [0.32.0]: https://github.com/dusk-network/phoenix/compare/v0.31.0...v0.32.0
 [0.31.0]: https://github.com/dusk-network/phoenix/compare/v0.30.0...v0.31.0